model: Extract `pycroft.model.unix_account`
